+++ test/integration/lib/testhelpers/helpers.go
@@ -2,6 +2,10 @@ package testhelpers
 
 import (
 	"context"
+	"fmt"
+	"os"
+	"os/exec"
+	"time"
 
 	"github.com/eclipse-symphony/symphony/test/integration/lib/shell"
 )
@@ -22,4 +26,47 @@ func CleanupManifests(ctx context.Context) error {
 		"kubectl delete targets.fabric.symphony --all -A",
 		"kubectl delete solutions.solution.symphony --all -A",
 	)
+}
+
+// Run a command with | or other things that do not work in shellcmd
+func ShellExec(cmd string) error {
+	fmt.Println("> ", cmd)
+
+	execCmd := exec.Command("sh", "-c", cmd)
+	execCmd.Stdout = os.Stdout
+	execCmd.Stderr = os.Stderr
+
+	return execCmd.Run()
+}
+
+// Run a mage command from /localenv
+func localenvCmd(mageCmd string, flavor string) error {
+	return ShellExec(fmt.Sprintf("cd ../../../localenv && mage %s %s", mageCmd, flavor))
+}
+
+// Prepare the cluster
+// Run this manually to prepare your local environment for testing/debugging
+func SetupCluster() error {
+	// Deploy symphony
+	err := localenvCmd("cluster:deploy", "")
+	if err != nil {
+		return err
+	}
+
+	// Wait a few secs for symphony cert to be ready;
+	// otherwise we will see error when creating symphony manifests in the cluster
+	// <Error from server (InternalError): error when creating
+	// "/mnt/vss/_work/1/s/test/integration/scenarios/basic/manifest/target.yaml":
+	// Internal error occurred: failed calling webhook "mtarget.kb.io": failed to
+	// call webhook: Post
+	// "https://symphony-webhook-service.default.svc:443/mutate-symphony-microsoft-com-v1-target?timeout=10s":
+	// x509: certificate signed by unknown authority>
+	time.Sleep(time.Second * 10)
+	return nil
+}
+
+// Clean up
+func Cleanup(testName string) {
+	localenvCmd(fmt.Sprintf("dumpSymphonyLogsForTest '%s'", testName), "")
+	localenvCmd("destroy all", "")
 }
