+++ test/integration/scenarios/00.unit/magefile.go
@@ -11,10 +11,9 @@ package main
 import (
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"time"
 
+	"github.com/eclipse-symphony/symphony/test/integration/lib/testhelpers"
 	"github.com/princjef/mageutil/shellcmd"
 )
 
@@ -37,7 +36,7 @@ func Test() error {
 
 	defer Cleanup()
 
-	err := Setup()
+	err := testhelpers.SetupCluster()
 	if err != nil {
 		return err
 	}
@@ -47,27 +46,6 @@ func Test() error {
 		return err
 	}
 
-	return nil
-}
-
-// Prepare the cluster
-// Run this manually to prepare your local environment for testing/debugging
-func Setup() error {
-	// Deploy symphony
-	err := localenvCmd("cluster:deploy", "")
-	if err != nil {
-		return err
-	}
-
-	// Wait a few secs for symphony cert to be ready;
-	// otherwise we will see error when creating symphony manifests in the cluster
-	// <Error from server (InternalError): error when creating
-	// "/mnt/vss/_work/1/s/test/integration/scenarios/basic/manifest/target.yaml":
-	// Internal error occurred: failed calling webhook "mtarget.kb.io": failed to
-	// call webhook: Post
-	// "https://symphony-webhook-service.default.svc:443/mutate-symphony-microsoft-com-v1-target?timeout=10s":
-	// x509: certificate signed by unknown authority>
-	time.Sleep(time.Second * 10)
 	return nil
 }
 
@@ -100,23 +78,6 @@ func Verify() error {
 
 // Clean up
 func Cleanup() {
-	shellExec(fmt.Sprintf("kubectl delete deployment nginx -n default"))
-	localenvCmd(fmt.Sprintf("dumpSymphonyLogsForTest '%s'", TEST_NAME), "")
-	localenvCmd("destroy all", "")
-}
-
-// Run a mage command from /localenv
-func localenvCmd(mageCmd string, flavor string) error {
-	return shellExec(fmt.Sprintf("cd ../../../localenv && mage %s %s", mageCmd, flavor))
-}
-
-// Run a command with | or other things that do not work in shellcmd
-func shellExec(cmd string) error {
-	fmt.Println("> ", cmd)
-
-	execCmd := exec.Command("sh", "-c", cmd)
-	execCmd.Stdout = os.Stdout
-	execCmd.Stderr = os.Stderr
-
-	return execCmd.Run()
+	testhelpers.ShellExec(fmt.Sprintf("kubectl delete deployment nginx -n default"))
+	testhelpers.Cleanup(TEST_NAME)
 }
