+++ test/integration/scenarios/02.basic/magefile.go
@@ -11,11 +11,11 @@ package main
 import (
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"time"
 
+	"github.com/eclipse-symphony/symphony/test/integration/lib/testhelpers"
 	"github.com/princjef/mageutil/shellcmd"
 )
 
@@ -53,8 +53,8 @@ var (
 func Test() error {
 	fmt.Println("Running ", TEST_NAME)
 
-	defer Cleanup()
-	err := SetupCluster()
+	defer testhelpers.Cleanup()
+	err := testhelpers.SetupCluster()
 	if err != nil {
 		return err
 	}
@@ -76,26 +76,6 @@ func Test() error {
 		time.Sleep(time.Second * 10)
 	}
 
-	return nil
-}
-
-// Prepare the cluster
-// Run this manually to prepare your local environment for testing/debugging
-func SetupCluster() error {
-	// Deploy symphony
-	err := localenvCmd("cluster:deploy", "")
-	if err != nil {
-		return err
-	}
-	// Wait a few secs for symphony cert to be ready;
-	// otherwise we will see error when creating symphony manifests in the cluster
-	// <Error from server (InternalError): error when creating
-	// "/mnt/vss/_work/1/s/test/integration/scenarios/basic/manifest/target.yaml":
-	// Internal error occurred: failed calling webhook "mtarget.kb.io": failed to
-	// call webhook: Post
-	// "https://symphony-webhook-service.default.svc:443/mutate-symphony-microsoft-com-v1-target?timeout=10s":
-	// x509: certificate signed by unknown authority>
-	time.Sleep(time.Second * 10)
 	return nil
 }
 
@@ -181,27 +161,6 @@ func CleanUpSymphonyObjects(namespace string) error {
 		return err
 	}
 	return nil
-}
-
-// Clean up
-func Cleanup() {
-	localenvCmd(fmt.Sprintf("dumpSymphonyLogsForTest '%s'", TEST_NAME), "")
-	localenvCmd("destroy all", "")
-}
-
-// Run a mage command from /localenv
-func localenvCmd(mageCmd string, flavor string) error {
-	return shellExec(fmt.Sprintf("cd ../../../localenv && mage %s %s", mageCmd, flavor))
-}
-
-func shellExec(cmd string) error {
-	fmt.Println("> ", cmd)
-
-	execCmd := exec.Command("sh", "-c", cmd)
-	execCmd.Stdout = os.Stdout
-	execCmd.Stderr = os.Stderr
-
-	return execCmd.Run()
 }
 
 func writeYamlStringsToFile(yamlString string, filePath string) error {
