+++ test/integration/scenarios/02.basicdelete/magefile.go
@@ -12,7 +12,6 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"time"
@@ -48,8 +47,8 @@ var (
 func Test() error {
 	fmt.Println("Running ", TEST_NAME)
 
-	defer Cleanup()
-	err := SetupCluster()
+	defer testhelpers.Cleanup(TEST_NAME)
+	err := testhelpers.SetupCluster()
 	if err != nil {
 		return err
 	}
@@ -78,26 +77,6 @@ func Test() error {
 		}
 	}
 
-	return nil
-}
-
-// Prepare the cluster
-// Run this manually to prepare your local environment for testing/debugging
-func SetupCluster() error {
-	// Deploy symphony
-	err := localenvCmd("cluster:deploy", "")
-	if err != nil {
-		return err
-	}
-	// Wait a few secs for symphony cert to be ready;
-	// otherwise we will see error when creating symphony manifests in the cluster
-	// <Error from server (InternalError): error when creating
-	// "/mnt/vss/_work/1/s/test/integration/scenarios/basic/manifest/target.yaml":
-	// Internal error occurred: failed calling webhook "mtarget.kb.io": failed to
-	// call webhook: Post
-	// "https://symphony-webhook-service.default.svc:443/mutate-symphony-microsoft-com-v1-target?timeout=10s":
-	// x509: certificate signed by unknown authority>
-	time.Sleep(time.Second * 10)
 	return nil
 }
 
@@ -199,19 +178,6 @@ func CleanUpSymphonyObjects(namespace string) error {
 	if err != nil {
 		return err
 	}
-
-	/*
-		instanceName := "instance02-v1"
-		solutionName := "solution02-v1"
-		err = shellcmd.Command(fmt.Sprintf("kubectl delete instances.solution.symphony %s -n %s", instanceName, namespace)).Run()
-		if err != nil {
-			return err
-		}
-
-		err = shellcmd.Command(fmt.Sprintf("kubectl delete solutions.solution.symphony %s -n %s", solutionName, namespace)).Run()
-		if err != nil {
-			return err
-		}*/
 	return nil
 }
 
@@ -264,27 +230,6 @@ func VerifyPodNotExists() error {
 	}
 
 	return nil
-}
-
-// Clean up
-func Cleanup() {
-	localenvCmd(fmt.Sprintf("dumpSymphonyLogsForTest '%s'", TEST_NAME), "")
-	localenvCmd("destroy all", "")
-}
-
-// Run a mage command from /localenv
-func localenvCmd(mageCmd string, flavor string) error {
-	return shellExec(fmt.Sprintf("cd ../../../localenv && mage %s %s", mageCmd, flavor))
-}
-
-func shellExec(cmd string) error {
-	fmt.Println("> ", cmd)
-
-	execCmd := exec.Command("sh", "-c", cmd)
-	execCmd.Stdout = os.Stdout
-	execCmd.Stderr = os.Stderr
-
-	return execCmd.Run()
 }
 
 func writeYamlStringsToFile(yamlString string, filePath string) error {
