+++ test/integration/scenarios/03.basicWithNsDelete/magefile.go
@@ -31,10 +31,9 @@ package main
 import (
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"time"
 
+	"github.com/eclipse-symphony/symphony/test/integration/lib/testhelpers"
 	"github.com/princjef/mageutil/shellcmd"
 )
 
@@ -66,7 +65,7 @@ var (
 func Test() error {
 	fmt.Println("Running ", TEST_NAME)
 
-	defer Cleanup()
+	defer testhelpers.Cleanup(TEST_NAME)
 
 	err := Setup()
 	if err != nil {
@@ -84,21 +83,8 @@ func Test() error {
 // Prepare the cluster
 // Run this manually to prepare your local environment for testing/debugging
 func Setup() error {
-	// Deploy symphony
-	err := localenvCmd("cluster:deploy", "")
-	if err != nil {
-		return err
-	}
+	testhelpers.SetupCluster()
 
-	// Wait a few secs for symphony cert to be ready;
-	// otherwise we will see error when creating symphony manifests in the cluster
-	// <Error from server (InternalError): error when creating
-	// "/mnt/vss/_work/1/s/test/integration/scenarios/basic/manifest/target.yaml":
-	// Internal error occurred: failed calling webhook "mtarget.kb.io": failed to
-	// call webhook: Post
-	// "https://symphony-webhook-service.default.svc:443/mutate-symphony-microsoft-com-v1-target?timeout=10s":
-	// x509: certificate signed by unknown authority>
-	time.Sleep(time.Second * 10)
 	// Deploy the manifests
 	for _, manifest := range testManifests {
 		fullPath, err := filepath.Abs(fmt.Sprintf(manifest, "oss"))
@@ -130,26 +116,4 @@ func Verify() error {
 	}
 
 	return nil
-}
-
-// Clean up
-func Cleanup() {
-	localenvCmd(fmt.Sprintf("dumpSymphonyLogsForTest '%s'", TEST_NAME), "")
-	localenvCmd("destroy all", "")
-}
-
-// Run a mage command from /localenv
-func localenvCmd(mageCmd string, flavor string) error {
-	return shellExec(fmt.Sprintf("cd ../../../localenv && mage %s %s", mageCmd, flavor))
-}
-
-// Run a command with | or other things that do not work in shellcmd
-func shellExec(cmd string) error {
-	fmt.Println("> ", cmd)
-
-	execCmd := exec.Command("sh", "-c", cmd)
-	execCmd.Stdout = os.Stdout
-	execCmd.Stderr = os.Stderr
-
-	return execCmd.Run()
 }
