+++ test/integration/scenarios/04.workflow/magefile.go
@@ -11,11 +11,11 @@ package main
 import (
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"time"
 
+	"github.com/eclipse-symphony/symphony/test/integration/lib/testhelpers"
 	"github.com/princjef/mageutil/shellcmd"
 	"gopkg.in/yaml.v2"
 )
@@ -78,7 +78,7 @@ func Test(labeling bool) error {
 		os.Setenv("labelingEnabled", "true")
 	}
 	defer Cleanup()
-	err := SetupCluster()
+	err := testhelpers.SetupCluster()
 	if err != nil {
 		return err
 	}
@@ -98,26 +98,6 @@ func Test(labeling bool) error {
 		}
 	}
 
-	return nil
-}
-
-// Prepare the cluster
-// Run this manually to prepare your local environment for testing/debugging
-func SetupCluster() error {
-	// Deploy symphony
-	err := localenvCmd("cluster:deploy", "")
-	if err != nil {
-		return err
-	}
-	// Wait a few secs for symphony cert to be ready;
-	// otherwise we will see error when creating symphony manifests in the cluster
-	// <Error from server (InternalError): error when creating
-	// "/mnt/vss/_work/1/s/test/integration/scenarios/basic/manifest/target.yaml":
-	// Internal error occurred: failed calling webhook "mtarget.kb.io": failed to
-	// call webhook: Post
-	// "https://symphony-webhook-service.default.svc:443/mutate-symphony-microsoft-com-v1-target?timeout=10s":
-	// x509: certificate signed by unknown authority>
-	time.Sleep(time.Second * 10)
 	return nil
 }
 
@@ -222,23 +202,7 @@ func Cleanup() {
 	if err != nil {
 		fmt.Printf("Failed to set up the symphony-ghcr-values.yaml. Please make sure the labelKey and labelValue is set to null.\n")
 	}
-	localenvCmd(fmt.Sprintf("dumpSymphonyLogsForTest '%s'", TEST_NAME), "")
-	localenvCmd("destroy all", "")
-}
-
-// Run a mage command from /localenv
-func localenvCmd(mageCmd string, flavor string) error {
-	return shellExec(fmt.Sprintf("cd ../../../localenv && mage %s %s", mageCmd, flavor))
-}
-
-func shellExec(cmd string) error {
-	fmt.Println("> ", cmd)
-
-	execCmd := exec.Command("sh", "-c", cmd)
-	execCmd.Stdout = os.Stdout
-	execCmd.Stderr = os.Stderr
-
-	return execCmd.Run()
+	testhelpers.Cleanup(TEST_NAME)
 }
 
 func writeYamlStringsToFile(yamlString string, filePath string) error {
