+++ test/integration/scenarios/05.catalog/magefile.go
@@ -14,10 +14,9 @@ import (
 	"flag"
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"time"
 
+	"github.com/eclipse-symphony/symphony/test/integration/lib/testhelpers"
 	"github.com/princjef/mageutil/shellcmd"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -63,9 +62,9 @@ var (
 func Test() error {
 	fmt.Println("Running ", TEST_NAME)
 
-	defer Cleanup()
+	defer testhelpers.Cleanup(TEST_NAME)
 
-	err := Setup()
+	err := testhelpers.SetupCluster()
 	if err != nil {
 		return err
 	}
@@ -75,27 +74,6 @@ func Test() error {
 		return err
 	}
 
-	return nil
-}
-
-// Prepare the cluster
-// Run this manually to prepare your local environment for testing/debugging
-func Setup() error {
-	// Deploy symphony
-	err := localenvCmd("cluster:deploy", "")
-	if err != nil {
-		return err
-	}
-
-	// Wait a few secs for symphony cert to be ready;
-	// otherwise we will see error when creating symphony manifests in the cluster
-	// <Error from server (InternalError): error when creating
-	// "/mnt/vss/_work/1/s/test/integration/scenarios/basic/manifest/target.yaml":
-	// Internal error occurred: failed calling webhook "mtarget.kb.io": failed to
-	// call webhook: Post
-	// "https://symphony-webhook-service.default.svc:443/mutate-symphony-microsoft-com-v1-target?timeout=10s":
-	// x509: certificate signed by unknown authority>
-	time.Sleep(time.Second * 10)
 	return nil
 }
 
@@ -226,26 +204,4 @@ func listCatalogs(namespace string, dynamicClient dynamic.Interface) (error, *un
 		return err, nil
 	}
 	return nil, catalogs
-}
-
-// Clean up
-func Cleanup() {
-	localenvCmd(fmt.Sprintf("dumpSymphonyLogsForTest '%s'", TEST_NAME), "")
-	localenvCmd("destroy all", "")
-}
-
-// Run a mage command from /localenv
-func localenvCmd(mageCmd string, flavor string) error {
-	return shellExec(fmt.Sprintf("cd ../../../localenv && mage %s %s", mageCmd, flavor))
-}
-
-// Run a command with | or other things that do not work in shellcmd
-func shellExec(cmd string) error {
-	fmt.Println("> ", cmd)
-
-	execCmd := exec.Command("sh", "-c", cmd)
-	execCmd.Stdout = os.Stdout
-	execCmd.Stderr = os.Stderr
-
-	return execCmd.Run()
 }
